refactor seed script to use modules
