+++ package.json
@@ -17,7 +17,7 @@
     "db:migrate": "drizzle-kit migrate --config drizzle.config.ts",
     "db:push": "drizzle-kit push --config drizzle.config.ts",
     "db:reset": "bun scripts/reset.ts",
-    "db:seed": "bun scripts/seed.ts",
+    "db:seed": "bun scripts/seed/seed.ts",
     "db:setup:local": "scripts/db-setup-local.sh",
     "db:generate:warehouse-schema": "bun scripts/generate-warehouse-schema.ts",
     "docker:kill": "scripts/docker-kill.sh",
