+++ scripts/seed/seed.ts
@@ -1,0 +1,9 @@
+import { seedRegions, seedWorkouts, enrichRegions } from './index';
+
+async function seedDatabase() {
+  await seedRegions();
+  await seedWorkouts();
+  await enrichRegions();
+}
+
+seedDatabase();
